Moved the integration tests out of the functions folder and versioned them. Need to locate an older version that works with v1bandit help
