+++ integration_tests/v2/pytest.ini
@@ -1,0 +1,4 @@
+[pytest]
+markers =
+    smoke_test: smoke tests (deselect with '-m "not smoke_test"')
+    pact_test: pact tests
