--- lambda_functions/v1/integration_tests/pytest.ini
@@ -1,4 +1,0 @@
-[pytest]
-markers =
-    smoke_test: smoke tests (deselect with '-m "not smoke_test"')
-    pact_test: pact tests
