--- lambda_functions/v2/directory_sha
@@ -1,1 +1,0 @@
-09b195270c96994
