--- lambda_functions/v2/integration_tests/pytest.ini
@@ -1,3 +1,0 @@
-[pytest]
-markers =
-    smoke_test: smoke tests (deselect with '-m "not smoke_test"')
